documentation strings added, properties separated
