+++ data_vis/operators/line_chart.py
@@ -1,3 +1,8 @@
+# File: line_chart.py
+# Author: Zdenek Dolezal
+# Licence: GPL 3.0
+# Description: Line chart implementation
+
 import bpy
 import math
 from mathutils import Vector
@@ -5,7 +10,8 @@ from mathutils import Vector
 
 from data_vis.utils.data_utils import find_data_range, find_axis_range, normalize_value, get_data_in_range
 from data_vis.operators.features.axis import AxisFactory
-from data_vis.general import OBJECT_OT_GenericChart, DV_LabelPropertyGroup, DV_AxisPropertyGroup, DV_HeaderPropertyGroup
+from data_vis.general import OBJECT_OT_GenericChart
+from data_vis.properties import DV_LabelPropertyGroup, DV_AxisPropertyGroup, DV_HeaderPropertyGroup
 from data_vis.data_manager import DataManager, DataType
 from data_vis.colors import NodeShader, ColorGen, ColorType
