+++ data_vis/operators/point_chart.py
@@ -1,8 +1,14 @@
+# File: point_chart.py
+# Author: Zdenek Dolezal
+# Licence: GPL 3.0
+# Description: Point chart (scatterplot) implementation
+
 import bpy
 from mathutils import Vector
 import math
 
-from data_vis.general import OBJECT_OT_GenericChart, DV_LabelPropertyGroup, DV_AxisPropertyGroup, DV_ColorPropertyGroup, DV_AnimationPropertyGroup, DV_HeaderPropertyGroup
+from data_vis.general import OBJECT_OT_GenericChart
+from data_vis.properties import DV_LabelPropertyGroup, DV_AxisPropertyGroup, DV_ColorPropertyGroup, DV_AnimationPropertyGroup, DV_HeaderPropertyGroup
 from data_vis.operators.features.axis import AxisFactory
 from data_vis.utils.data_utils import normalize_value
 from data_vis.colors import ColoringFactory, ColorType
